Moved classes to root namespace
