+++ src/Auth/AuthConfig.php
@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace Tempest\Auth;
 
-use Tempest\Auth\Authenticators\DatabaseAuthenticator;
-use Tempest\Auth\Contracts\Authenticator;
-
 final class AuthConfig
 {
     public function __construct(
