+++ src/Auth/GenericAuthenticator.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Tempest\Auth\Authenticators;
+namespace Tempest\Auth;
 
-use Tempest\Auth\Contracts\Authenticator;
 use Tempest\Http\Session\Session;
 
 abstract class GenericAuthenticator implements Authenticator
